`MainActivity` 内Fragmentの命名fix
