+++ app/src/main/java/sugtao4423/lod/ui/main/fragment/HomeFragment.kt
@@ -17,14 +17,14 @@ import sugtao4423.lod.databinding.FragmentListBinding
 import sugtao4423.lod.tweetlistview.EndlessScrollListener
 import sugtao4423.lod.tweetlistview.TweetListAdapter
 import twitter4j.Paging
-import twitter4j.ResponseList
 import twitter4j.Status
 import twitter4j.TwitterException
 
-class Fragment_mention : Fragment() {
+class HomeFragment : Fragment() {
 
     private lateinit var binding: FragmentListBinding
     private lateinit var adapter: TweetListAdapter
+    private var listAsTL = -1L
     private lateinit var app: App
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
@@ -36,10 +36,10 @@ class Fragment_mention : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         app = requireContext().applicationContext as App
+        listAsTL = app.account.listAsTL
 
         adapter = TweetListAdapter(requireContext())
         binding.listLine.adapter = adapter
-
         val scrollListener = getLoadMoreListener()
         binding.listLine.addOnScrollListener(scrollListener)
 
@@ -47,24 +47,14 @@ class Fragment_mention : Fragment() {
                 android.R.color.holo_orange_light, android.R.color.holo_red_light)
         val onRefreshListener = SwipeRefreshLayout.OnRefreshListener {
             adapter.clear()
-            loadMention()
+            loadTimeLine()
             scrollListener.resetState()
         }
         binding.listPull2Refresh.setOnRefreshListener(onRefreshListener)
         onRefreshListener.onRefresh()
     }
 
-    private fun getLoadMoreListener(): EndlessScrollListener {
-        return object : EndlessScrollListener(binding.listLine.linearLayoutManager) {
-            override fun onLoadMore(currentPage: Int) {
-                if (adapter.hasNextPage) {
-                    loadMention()
-                }
-            }
-        }
-    }
-
-    private fun loadMention() {
+    private fun loadTimeLine() {
         CoroutineScope(Dispatchers.Main).launch {
             val result = withContext(Dispatchers.IO) {
                 val paging = Paging(1, 50).let {
@@ -72,7 +62,11 @@ class Fragment_mention : Fragment() {
                 }
 
                 try {
-                    app.twitter.getMentionsTimeline(paging)
+                    if (listAsTL > 0) {
+                        app.twitter.getUserListStatuses(listAsTL, paging)
+                    } else {
+                        app.twitter.getHomeTimeline(paging)
+                    }
                 } catch (e: TwitterException) {
                     null
                 }
@@ -81,9 +75,12 @@ class Fragment_mention : Fragment() {
                 if (result.isEmpty()) {
                     adapter.hasNextPage = false
                 }
-                addAll(result)
+                if (adapter.itemCount <= 0) {
+                    app.latestTweetId = result[0].id
+                }
+                adapter.addAll(result)
             } else {
-                ShowToast(requireContext().applicationContext, R.string.error_get_mention)
+                ShowToast(requireContext().applicationContext, R.string.error_get_timeline)
             }
             binding.listPull2Refresh.isRefreshing = false
             binding.listPull2Refresh.isEnabled = true
@@ -92,13 +89,20 @@ class Fragment_mention : Fragment() {
 
     fun insert(status: Status) {
         adapter.insertTop(status)
+        app.latestTweetId = status.id
         if (binding.listLine.linearLayoutManager.findFirstVisibleItemPosition() <= 1) {
             binding.listLine.smoothScrollToPosition(0)
         }
     }
 
-    private fun addAll(statuses: ResponseList<Status>) {
-        adapter.addAll(statuses)
+    private fun getLoadMoreListener(): EndlessScrollListener {
+        return object : EndlessScrollListener(binding.listLine.linearLayoutManager) {
+            override fun onLoadMore(currentPage: Int) {
+                if (adapter.hasNextPage) {
+                    loadTimeLine()
+                }
+            }
+        }
     }
 
 }
