+++ app/src/main/res/layout/fragment_list.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="viewModel"
-            type="sugtao4423.lod.ui.main.fragment.Fragment_ListViewModel" />
+            type="sugtao4423.lod.ui.main.fragment.ListFragmentViewModel" />
     </data>
 
     <RelativeLayout
